Conflicts:
	gr-fec/examples/CMakeLists.txt
	gr-fec/examples/ber_curve_gen.grc
	gr-fec/lib/CMakeLists.txt
	gr-fec/swig/fec_swig.i

Merging upstream and also merging in my ldpc_mods branch into master.

+++ gr-fec/lib/async_decoder_impl.h
@@ -44,7 +44,7 @@ namespace gr {
       bool d_packed;
       bool d_rev_pack;
 
-      int d_max_bits_in;
+      size_t d_max_bits_in;
       float *d_tmp_f32;
       int8_t *d_tmp_u8;
       uint8_t *d_bits_out;
